splitting index.js into separate modules
