+++ django_tutorial/first_project/web_app/templates/web_app/index.html
@@ -48,8 +48,6 @@
                 integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4="
                 crossorigin="anonymous"></script>
         <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.js"></script>
-        <script src="https://unpkg.com/redux@3.7.2/dist/redux.min.js"></script>
-        <script src="https://unpkg.com/redux-thunk@2.1.0/dist/redux-thunk.min.js"></script>
         <script src="{% static "web_app/js/index.bundle.js" %}"></script>
         <script async defer
                 src="https://maps.googleapis.com/maps/api/js?key=AIzaSyCMkwqz_xwd2fA2nAKy5s5ruzdAIy6T354&callback=init">
